move OpenGL Framedebugger GSH to Qt
